@@ -11,22 +11,16 @@
   "license": "ISC",
   "dependencies": {
     "babel-core": "6.26.0",
-    "babel-plugin-transform-async-functions": "6.22.0",
-    "babel-plugin-transform-builtin-extend": "1.1.2",
-    "babel-plugin-transform-decorators-legacy": "1.3.4",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-es2016": "6.24.1",
-    "babel-preset-es2017": "6.24.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-2": "6.24.1",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
     "babel-runtime": "6.26.0",
     "bootstrap": "4.0.0",
     "classnames": "2.2.5",
     "enzyme": "3.3.0",
-    "enzyme-adapter-react-15.4": "1.0.3",
+    "enzyme-adapter-react-16": "^1.1.1",
     "jest": "22.4.3",
-    "react": "15.4.2",
-    "react-addons-test-utils": "=15.4.2",
-    "react-dom": "15.4.2"
+    "prop-types": "^15.6.1",
+    "react": "^16.3.1",
+    "react-dom": "^16.3.1"
   }
 }
