@@ -1,13 +1,12 @@
-import assert from 'assert';
 import React from 'react';
 import { shallow } from '../tests/helpers';
 import { Alert, AlertType } from './alert.jsx';
 
-describe('ButtonGroup', () => {
+describe('Alert', () => {
     it('large size', () => {
         const wrapper = shallow(<Alert className="foo" type={AlertType.Info}><span>bar</span></Alert>);
 
-        assert.ok(wrapper.contains(<span>bar</span>));
-        assert.equal(wrapper.prop('className'), 'foo alert alert-info');
+        expect(wrapper.contains(<span>bar</span>));
+        expect(wrapper.prop('className')).toEqual('foo alert alert-info');
     });
 });
